chore: adjust support directory naming
