+++ support/middleware/authentication.go
@@ -4,9 +4,9 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/zetsux/gin-gorm-clean-starter/common/base"
 	"github.com/zetsux/gin-gorm-clean-starter/core/helper/messages"
 	"github.com/zetsux/gin-gorm-clean-starter/core/service"
+	"github.com/zetsux/gin-gorm-clean-starter/support/base"
 
 	"github.com/gin-gonic/gin"
 )
