+++ support/middleware/authorization.go
@@ -5,9 +5,9 @@ import (
 	"slices"
 
 	"github.com/gin-gonic/gin"
-	"github.com/zetsux/gin-gorm-clean-starter/common/base"
-	"github.com/zetsux/gin-gorm-clean-starter/common/constant"
 	"github.com/zetsux/gin-gorm-clean-starter/core/helper/messages"
+	"github.com/zetsux/gin-gorm-clean-starter/support/base"
+	"github.com/zetsux/gin-gorm-clean-starter/support/constant"
 )
 
 func Authorize(roles ...string) gin.HandlerFunc {
