+++ support/middleware/cors.go
