+++ support/middleware/tests/authentication_test.go
@@ -7,20 +7,20 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/require"
-	"github.com/zetsux/gin-gorm-clean-starter/common/middleware"
 	"github.com/zetsux/gin-gorm-clean-starter/core/service"
+	"github.com/zetsux/gin-gorm-clean-starter/support/middleware"
 )
 
 // --- Test Helpers ---
 
-func setupAuthorizationTest(t *testing.T) (*gin.Engine, service.JWTService) {
+func setupAuthenticationTest(t *testing.T) (*gin.Engine, service.JWTService) {
 	t.Parallel()
 	gin.SetMode(gin.TestMode)
 	r := gin.New()
 	r.Use(middleware.ErrorHandler())
 
 	jwtS := service.NewJWTService()
-	r.GET("/protected", middleware.Authenticate(jwtS), middleware.Authorize(), func(c *gin.Context) {
+	r.GET("/protected", middleware.Authenticate(jwtS), func(c *gin.Context) {
 		c.String(http.StatusOK, "ok")
 	})
 
@@ -29,28 +29,22 @@ func setupAuthorizationTest(t *testing.T) (*gin.Engine, service.JWTService) {
 
 // --- Tests ---
 
-func TestAuthorize_ForbiddenRole(t *testing.T) {
-	r, jwtS := setupAuthorizationTest(t)
+func TestAuthenticate_MissingToken(t *testing.T) {
+	r, _ := setupAuthenticationTest(t)
+
+	req := httptest.NewRequest(http.MethodGet, "/protected", nil)
+	w := httptest.NewRecorder()
+	r.ServeHTTP(w, req)
+	require.Equal(t, http.StatusUnauthorized, w.Code)
+}
+
+func TestAuthenticate_ValidTokenAndRole(t *testing.T) {
+	r, jwtS := setupAuthenticationTest(t)
 
 	token := jwtS.GenerateToken("abc", "user")
 	req := httptest.NewRequest(http.MethodGet, "/protected", nil)
 	req.Header.Set("Authorization", "Bearer "+token)
 	w := httptest.NewRecorder()
 	r.ServeHTTP(w, req)
-	require.Equal(t, http.StatusForbidden, w.Code)
-}
-
-func TestAuthorize_ValidRole(t *testing.T) {
-	r, jwtS := setupAuthorizationTest(t)
-
-	token := jwtS.GenerateToken("abc", "admin")
-
-	req := httptest.NewRequest(http.MethodGet, "/protected", nil)
-	req.Header.Set("Authorization", "Bearer "+token)
-	w := httptest.NewRecorder()
-
-	r.ServeHTTP(w, req)
-
 	require.Equal(t, http.StatusOK, w.Code)
-	require.Equal(t, "ok", w.Body.String())
 }
