+++ tests/integration/user/admin_test.go
@@ -8,9 +8,9 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/zetsux/gin-gorm-clean-starter/common/base"
-	"github.com/zetsux/gin-gorm-clean-starter/common/constant"
 	"github.com/zetsux/gin-gorm-clean-starter/core/helper/dto"
+	"github.com/zetsux/gin-gorm-clean-starter/support/base"
+	"github.com/zetsux/gin-gorm-clean-starter/support/constant"
 	"github.com/zetsux/gin-gorm-clean-starter/tests/testutil"
 	"github.com/zetsux/gin-gorm-clean-starter/tests/testutil/factory"
 )
